Float撲滅した!!(一部のこってるけどねｗ)

+++ Client/src/main/java/com/mc1510ty/LWJGLVoxelGame/Client/Button.java
@@ -16,10 +16,10 @@
 package com.mc1510ty.LWJGLVoxelGame.Client;
 
 public class Button {
-    private float x, y, width, height;
+    private double x, y, width, height;
     private String label;
 
-    public Button(float x, float y, float width, float height, String label) {
+    public Button(double x, double y, double width, double height, String label) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -31,9 +31,9 @@ public class Button {
         return mouseX >= x && mouseX <= x + width && mouseY >= y && mouseY <= y + height;
     }
 
-    public float getX() { return x; }
-    public float getY() { return y; }
-    public float getWidth() { return width; }
-    public float getHeight() { return height; }
+    public double getX() { return x; }
+    public double getY() { return y; }
+    public double getWidth() { return width; }
+    public double getHeight() { return height; }
     public String getLabel() { return label; }
 }
